requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.6.0
python-dotenv==1.0.1
Pillow==10.4.0
boto3==1.35.36
celery==5.3.6
redis==5.0.6
